pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.0",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.30.0",
    "bcrypt>=5.0.0",
    "fastapi~=0.118.2",
    "ldap3>=2.10.2",
    "passlib>=1.7.4",
    "pydantic~=2.12.0",
    "pydantic-settings>=2.11.0",
    "python-jose>=3.5.0",
    "sqlalchemy[asyncio]>=2.0.44",
    "uuid6>=2025.0.1",
    "uvicorn~=0.37.0",
    "pillow>=11.1.0",
]
